# apps/server/openapi.yml
openapi: 3.0.0
paths:
...
  /api/auth/register:
    post:
      description: >-
        Complete user registration after OAuth. Requires pending-registration
        cookie.
      operationId: AuthController_register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: No pending registration found or invalid input
      summary: Register a new user
      tags: *ref_4
...
  /api/auth/refresh:
    post:
      description: Get a new access token using the refresh token cookie
      operationId: AuthController_refresh
      parameters: []
      responses:
        '200':
          description: Access token refreshed successfully
        '401':
          description: No refresh token provided or invalid/expired token
      security:
        - refresh-token: []
      summary: Refresh access token
      tags: *ref_4
...
  /api/auth/me:
    get:
      description: Get the authenticated user information
      operationId: AuthController_getCurrentUser
      parameters: []
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Not authenticated
      security:
        - bearer: []
      summary: Get current user
      tags: *ref_4
...
  /api/user:
    get:
      description: Returns the authenticated user information
      operationId: UserController_getCurrentUser
      parameters: []
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '401':
          description: Not authenticated
      security: &ref_6
        - bearer: []
      summary: Get current user
      tags: &ref_7
        - User
    put:
      description: Update the authenticated user profile
      operationId: UserController_updateUser
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequestDto'
      responses:
        '200':
          description: Updated user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '400':
          description: Invalid input data
        '401':
          description: Not authenticated
      security: *ref_6
      summary: Update current user
      tags: *ref_7
info:
  title: RealWorld API
  description: Conduit API specification - RealWorld example app
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    RegisterDto:
      type: object
      properties:
        username:
          type: string
          description: Username (alphanumeric and underscore only, min 3 characters)
          example: john_doe
          minLength: 3
      required:
        - username
    RegisterResponse:
      type: object
      properties:
        id:
          type: string
          description: User public ID
          example: clxxxxxxxxxxxxxxxxxx
        username:
          type: string
          description: Username
          example: john_doe
        email:
          type: string
          description: Email address
          example: john@example.com
      required:
        - id
        - username
        - email
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          description: User public ID
          example: clxxxxxxxxxxxxxxxxxx
        username:
          type: string
          description: Username
          example: john_doe
      required:
        - id
        - username
...
    UserDto:
      type: object
      properties:
        email:
          type: string
          description: Email address
          example: user@example.com
        username:
          type: string
          description: Username
          example: john_doe
        bio:
          type: object
          description: User bio
          example: I like to code
          nullable: true
        image:
          type: object
          description: User image URL
          example: https://example.com/avatar.jpg
          nullable: true
      required:
        - email
        - username
        - bio
        - image
    UserResponseDto:
      type: object
      properties:
        user:
          description: User object
          allOf:
            - $ref: '#/components/schemas/UserDto'
      required:
        - user
    UpdateUserDto:
      type: object
      properties:
        email:
          type: string
          description: Email address
          example: newemail@example.com
        username:
          type: string
          description: Username
          example: new_username
          minLength: 3
        password:
          type: string
          description: Password (not used in OAuth mode)
        bio:
          type: string
          description: User bio
          example: Updated bio
        image:
          type: string
          description: User image URL
          example: https://example.com/new-avatar.jpg
    UpdateUserRequestDto:
      type: object
      properties:
        user:
          description: User update data
          allOf:
            - $ref: '#/components/schemas/UpdateUserDto'
      required:
        - user